reduced duplicated login logic
